server/requirements/base.txt
Django==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg[binary]==3.2.12
gunicorn==23.0.0
whitenoise==6.9.0
django-storages[s3]==1.14.6
reportlab==4.4.1
